refactor: move policies, commands and events into domain
