+++ src/app/domain/policies/whenever-comment-created-invalidate-cache.ts
@@ -1,6 +1,6 @@
-import type { Comment } from "../domain/entities/comment.js";
-import type { CacheStore } from "../driven-ports/cache-store.js";
-import type { EventBroker } from "../driven-ports/event-broker.js";
+import type { Comment } from "../entities/comment.js";
+import type { CacheStore } from "../../driven-ports/cache-store.js";
+import type { EventBroker } from "../../driven-ports/event-broker.js";
 
 type WheneverCommentCreatedInvalidateCachePolicy = ({
   eventBroker,
