+++ src/app/index.ts
@@ -2,10 +2,10 @@ import type { DataStore } from "./driven-ports/data-store.js";
 import type { App } from "./domain/entities/app.js";
 import type { EventBroker } from "./driven-ports/event-broker.js";
 import type { RandomId } from "./driven-ports/random-id.js";
-import { commandCreateComment } from "./commands/create-comment/index.js";
-import { commandUpdateComment } from "./commands/update-comment/index.js";
-import { commandDeleteComment } from "./commands/delete-comment/index.js";
-import { queryGetCommentsForHost } from "./queries/get-comments-for-host/index.js";
+import { commandCreateComment } from "./domain/commands/create-comment/index.js";
+import { commandUpdateComment } from "./domain/commands/update-comment/index.js";
+import { commandDeleteComment } from "./domain/commands/delete-comment/index.js";
+import { queryGetCommentsForHost } from "./domain/queries/get-comments-for-host/index.js";
 import { toCommentCreatedEvent } from "./domain/helpers/events/comment-created.js";
 import { toCommentUpdatedEvent } from "./domain/helpers/events/comment-updated.js";
 import { toCommentDeletedEvent } from "./domain/helpers/events/comment-deleted.js";
